@@ -37,9 +37,9 @@
     "query-string": "^6.2.0",
     "raf": "3.4.0",
     "ramda": "^0.25.0",
-    "react": "^16.6.3",
+    "react": "^16.7.0-alpha.2",
     "react-dev-utils": "^5.0.2",
-    "react-dom": "^16.6.3",
+    "react-dom": "^16.7.0-alpha.2",
     "react-router-dom": "^4.3.1",
     "react-select": "^2.1.2",
     "react-virtualized": "^9.21.0",
