@@ -1,4 +1,4 @@
-import React, { Component } from 'react'
+import React, { Component, useState } from 'react'
 import PlaceSelector from 'components/PlaceSelector'
 import P from 'prop-types'
 import {
@@ -11,117 +11,114 @@ import {
   ItemActions,
 } from './style.js'
 
-class DeviceItem extends Component {
-  constructor(props) {
-    super(props)
-    const { name, ip, description, place } = props
-    this.state = {
-      editState: false,
-      curName: name,
-      curIp: ip,
-      curPlace: place,
-      curDescription: description,
-    }
+function DeviceItem(props) {
+  const {
+    name,
+    ip,
+    description,
+    place,
+    saveChanges,
+    active,
+    onDelete,
+    handleChange,
+    index,
+    style,
+  } = props
+  const [editState, setEditState] = useState(false)
+  const [curName, setCurName] = useState(name)
+  const [curIp, setCurIp] = useState(ip)
+  const [curPlace, setCurPlace] = useState(place)
+  const [curDescription, setCurDescription] = useState(description)
+
+  const toggleEditState = () => {
+    setEditState(!editState)
   }
 
-  onSave = () => {
-    const { saveChanges } = this.props
-    const {
-      curDescription: description,
-      curIp: ip,
-      curName: name,
-      curPlace: place,
-    } = this.state
-    const item = { description, ip, name, place }
+  const onSave = () => {
+    const item = { curDescription, curIp, curName, curPlace }
+    console.table(item)
     saveChanges(item)
-    this.toggleEditState()
-  }
-  toggleEditState = () => {
-    const { editState } = this.state
-    this.setState({ editState: !editState })
+    toggleEditState()
   }
 
-  render() {
-    const { active, onDelete, handleChange, index, style } = this.props
-    const { editState, curPlace, curName, curIp, curDescription } = this.state
-    return (
-      <Row style={style}>
-        <IdItem>
-          <input type="checkbox" onChange={handleChange} checked={active} />
-          {index}
-        </IdItem>
-        <ItemName>
-          {editState ? (
-            <input
-              type="text"
-              value={curName}
-              onChange={e => {
-                this.setState({ curName: e.target.value })
-              }}
-            />
-          ) : (
-            curName
-          )}
-        </ItemName>
-        <ItemIP>
-          {editState ? (
-            <input
-              type="text"
-              value={curIp}
-              onChange={e => {
-                this.setState({ curIp: e.target.value })
-              }}
-            />
-          ) : (
-            curIp
-          )}
-        </ItemIP>
-        <ItemDescription>
-          {editState ? (
-            <input
-              type="text"
-              value={curDescription}
-              onChange={e => {
-                this.setState({ curDescription: e.target.value })
-              }}
-            />
-          ) : (
-            curDescription
-          )}
-        </ItemDescription>
-        <ItemPlaceSelector>
-          {editState ? (
-            <PlaceSelector
-              value={curPlace.id}
-              onChange={e => {
-                this.setState({ curPlace: e })
-              }}
-            />
-          ) : (
-            curPlace.name
-          )}
-        </ItemPlaceSelector>
-        <ItemActions>
-          <button type="button" onClick={onDelete}>
+  return (
+    <Row style={style}>
+      <IdItem>
+        <input type="checkbox" onChange={handleChange} checked={active} />
+        {index}
+      </IdItem>
+      <ItemName>
+        {editState ? (
+          <input
+            type="text"
+            value={curName}
+            onChange={e => {
+              setCurName(e.target.value)
+            }}
+          />
+        ) : (
+          curName
+        )}
+      </ItemName>
+      <ItemIP>
+        {editState ? (
+          <input
+            type="text"
+            value={curIp}
+            onChange={e => {
+              setCurIp(e.target.value)
+            }}
+          />
+        ) : (
+          curIp
+        )}
+      </ItemIP>
+      <ItemDescription>
+        {editState ? (
+          <input
+            type="text"
+            value={curDescription}
+            onChange={e => {
+              setCurDescription(e.target.value)
+            }}
+          />
+        ) : (
+          curDescription
+        )}
+      </ItemDescription>
+      <ItemPlaceSelector>
+        {editState ? (
+          <PlaceSelector
+            value={curPlace.id}
+            onChange={e => {
+              setCurPlace(e)
+            }}
+          />
+        ) : (
+          curPlace.name
+        )}
+      </ItemPlaceSelector>
+      <ItemActions>
+        <button type="button" onClick={onDelete}>
+          {' '}
+          delete{' '}
+        </button>
+        {editState ? (
+          <button type="button" onClick={onSave}>
             {' '}
-            delete{' '}
+            save{' '}
           </button>
-          {editState ? (
-            <button type="button" onClick={this.onSave}>
-              {' '}
-              save{' '}
-            </button>
-          ) : (
-            <button type="button" onClick={this.toggleEditState}>
-              {' '}
-              edit{' '}
-            </button>
-          )}
-        </ItemActions>
-      </Row>
-    )
-  }
+        ) : (
+          <button type="button" onClick={toggleEditState}>
+            {' '}
+            edit{' '}
+          </button>
+        )}
+      </ItemActions>
+    </Row>
+  )
 }
+
 DeviceItem.defaultProps = {
   place: {},
   name: '',
